requirements.txt
PyPDF2
streamlit
streamlit-chat
langchain
langchain-community
langchain-core
langchain-openai
chromadb
sentence-transformers
python-dotenv